feat: Add Menu.html template for displaying menus
This commit adds the Menu.html template, which is used to display menus for different places. The template extends the uni_base.html template and includes a university-container div that contains university-box divs for each menu item. Each university-box displays the product name, description, and price. Additionally, an "Add to Favorites" button is included at the bottom of the page. This template is intended to be used in conjunction with the Places page and the lazy loading feature for images.