feat: Add form to set place as favorite in Menu.html
This commit adds a form to the Menu.html template that allows users to set a place as their favorite. The form includes a button that triggers a POST request to the '/set_favorite' endpoint with the place ID as a parameter. This feature enhances the user experience by providing an easy way to add places to their favorites list.