feat: Add link to list of universities in home.html
This commit adds a link to the list of universities in the home.html template. The link is displayed as a button and directs users to the '/universities' endpoint. This addition enhances the user experience by providing easy access to the list of universities, allowing users to explore and discover new places near different universities.