feat: Update nav_base_admin.html to fix link to universities
The link in nav_base_admin.html has been updated to point to the correct URL for universities. Previously, it was pointing to "/places" but it has been changed to "/universities" to accurately reflect the purpose of the page. This change ensures that users are directed to the correct page when navigating through the admin panel.